Revert "added default img, dashboard redirection, categories, news filtering"
This reverts commit 209fc61750b938d8c0c6f7cade1a64099ae973e5.

+++ app/Http/Controllers/PageController.php
@@ -3,137 +3,61 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Storage;
-use App\Category;
 use App\Page;
 use App\File;
-use App\Comment;
-use Purifier;
 
 class PageController extends Controller
 {
 
-    public function __construct()
-    {
+    public function __construct(){
         $this->middleware('auth', ['except'=> ['index', 'show']]);
     }
 
+    public function index() {
+      //$pages = Page::orderBy('created_at','desc')->paginate(4);
+        $pages = Page::latest()->paginate(4);
 
-
-    public function index()
-    {
-        $pages = Page::with('categories')->orderBy('created_at','desc')
-                                         ->paginate(4);
-
-        return view('page.index', ['pages' => $pages]);
+        return view('page.index', ['pages'=>$pages]);
     }
 
-
-
-
-
-    public function show($id)
-    {
+    public function show($id){
         $page = Page::find($id);
-
         return view('page.show', ['page' => $page]);
     }
 
 
-
-    /* CREATE AND STORE */
-    public function create()
-    {
-        $categories = Category::get()->pluck('category','id');
-
-        return view('page.create', ['categories' => $categories]);
-    }
-
-
-
-    public function store( Request $request )
-    {
-        $this->validate( $request, [
-            'title' => 'required',
-            'content' => 'required',
-            'name' => 'image|nullable|max:1999',
-            'category.*' => 'exists:category,id'
-        ]);
-
-
-        $page = new Page;
-        $page->title = $request->input('title');
-        $page->content = Purifier::clean($request->input('content'));
-        $page->user_id = Auth()->user()->id;
-        $page->save();
-
-        //add categories
-        $page->categories()->sync((array)$request->input('category'));
-
-
-        //create new file
-        $file = new File;
-
-        if($request->hasFile('name')){
-            $filenameWithExtension = $request->file('name')->getClientOriginalName();
-            $filename = pathinfo($filenameWithExtension, PATHINFO_FILENAME);
-            $extension = $request->file('name')->getClientOriginalExtension();
-            $filenameToStore = $filename.'_'.time().'.'.$extension;
-            $path = $request->file('name')->storeAs('public/images', $filenameToStore);
-        }else{
-            $filenameToStore = 'defaultimg.jpg';
-        }
-
-        $file->name=$filenameToStore;
-        $file->page_id = $page->id;
-        $file->save();
-
-
-        return redirect('/')->with('success', 'Page created successfuly.');
-    }
-
-
-
     /* EDIT and UPDATE */
-    public function edit($id)
-    {
-       $page = Page::findOrFail($id);
-       $categories = Category::get()->pluck('category', 'id');
+    public function edit($id) {
+       $page = Page::find($id);
 
        if(Auth()->user()->id != $page->user_id){
-           return redirect('page')->with('error', ' You have no permission for this action.');
+           return redirect('page')->with('error', 'you have no permission for this action');
        }
 
-       return view('page.edit',  ['page' => $page, 'categories' => $categories]);
+       return view('page.edit', ['page' => $page]);
     }
 
 
+        public function update(Request $request, $id) {
 
-
-
-        public function update(Request $request, $id)
-        {
             $this->validate($request, [
-                'title'=>'required',
-                'content'=> 'required',
-                'name' => 'image|nullable|max:1999',
-                'category.*' => 'exists:category,id'
+            'title'=>'required',
+            'content'=> 'required',
+            'name' => 'image|nullable|max:1999'
             ]);
 
             $page =  Page::find($id);
             $page->title = $request->input('title');
-            $page->content = Purifier::clean($request->input('content'));
+            $page->content = $request->input('content');
             $page->save();
 
 
-            $page->categories()->sync((array)$request->input('category'));
-
-
-
+            //File upload
+            // !! ??
             $file = File::where('page_id', '=', $page->id)->first();
+            if($file == null){ $file = new File;}
 
             if($request->hasFile('name')){
-
                 $filenameWithExtension = $request->file('name')->getClientOriginalName();
                 $filename = pathinfo($filenameWithExtension, PATHINFO_FILENAME);
                 $extension = $request->file('name')->getClientOriginalExtension();
@@ -145,40 +69,61 @@ class PageController extends Controller
                 $file->save();
             }
 
-           return redirect('dashboard')->with('success', 'Page updated successfuly.');
+           return redirect('/page')->with('success', 'Page updated');
         }
 
 
 
+    /* CREATE AND STORE */
+    public function create(){
+        return view('page.create');
+    }
 
+    public function store( Request $request ) {
 
+        $this->validate( $request, [
+            'title' => 'required',
+            'content' => 'required',
+            'name' => 'image|nullable|max:1999'
+        ]);
 
-    public function destroy($id)
-    {
+        $page = new Page;
+        $page->title = $request->input('title');
+        $page->content = $request->input('content');
+        $page->user_id = Auth()->user()->id;
+        $page->save();
+
+        //File upload  ??!!
+        $file = new File;
+
+        if($request->hasFile('name')){
+            $filenameWithExtension = $request->file('name')->getClientOriginalName();
+            $filename = pathinfo($filenameWithExtension, PATHINFO_FILENAME);
+            $extension = $request->file('name')->getClientOriginalExtension();
+            $filenameToStore = $filename.'_'.time().'.'.$extension;
+            $path = $request->file('name')->storeAs('public/images', $filenameToStore);
+        }
+
+        $file->name=$filenameToStore;
+        $file->page_id = $page->id;
+        $file->save();
+
+        return redirect('/page')->with('success', 'Page created.');
+    }
+
+    public function destroy($id) {
         $page = Page::find($id);
 
         if(Auth()->user()->id != $page->user_id){
             return redirect('page')->with('error', 'You have no permission for this action!');
         }
 
-
-        $page->categories()->detach();
-
-        //Delete file model and file from storage (except default image)
-        $file = File::where('page_id', $page->id)->first();
-        if ($file->name != 'defaultimg.jpg')  {
-            $file->delete();
-            Storage::delete('public/images/'. $file->name);
-        }else{
-            $file->delete();
-        }
-            
-        $comments = Comment::where('page_id', $page->id);
-        if ($comments != null)  {   $comments->delete();  }
+        // ?? !!
+        $file = File::where('page_id', '=', $page->id)->first();
+        if ($file != null)  {   $file->delete();  }
 
         $page->delete();
-
-        return redirect('dashboard')->with('success', 'Page removed successfuly.');
+        return redirect('/page')->with('success', 'Page removed.');
     }
